requirements.txt
pysnow==0.7.17
langchain==0.1.0
langchain-nvidia-ai-endpoints==0.1.0
openai==1.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
flask==3.0.0
pyngrok==5.2.0
werkzeug==3.0.0
